+++ EliteLib/Structs/stock_item.cs
@@ -4,14 +4,14 @@
     {
         internal string name;
         internal int current_quantity;
-        internal int current_price;
-        internal int base_price;
+        internal float current_price;
+        internal float base_price;
         internal int eco_adjust;
         internal int base_quantity;
         internal int mask;
         internal string units;
 
-        internal stock_item(string name, int current_quantity, int current_price, int base_price, int eco_adjust, int base_quantity, int mask, string units)
+        internal stock_item(string name, int current_quantity, float current_price, float base_price, int eco_adjust, int base_quantity, int mask, string units)
         {
             this.name = name;
             this.current_quantity = current_quantity;
